% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_WB_c.R
\name{scale_fill_WB_c}
\alias{scale_fill_WB_c}
\title{Use WrensBookshelf with ggplot}
\usage{
scale_fill_WB_c(name, direction = 1, ...)
}
\arguments{
\item{name}{Name of palette from WrensBookshelf.}

\item{direction}{Direction you want the palette to proceed; 1 = direction shown in ShowBookshelf() or ShowBook(), -1 = reversed.}

\item{...}{See scale_color_gradientn() for additional params.}
}
\value{

}
\description{
Wrapper function to allow easy integration with ggplot2. Sets continuous fill scale.
}
\examples{
ggplot(mtcars, aes(x = disp,y = wt, fill= mpg))+geom_point(size=3, shape = 21)+scale_color_WB_c(name = "What We'll Build")
}
