% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowBookshelf.R
\name{ShowBookshelf}
\alias{ShowBookshelf}
\title{View all of Wren's Books}
\usage{
ShowBookshelf(CVDsafe = FALSE, BestFor = NA)
}
\arguments{
\item{CVDsafe}{Argument used to restrict displayed palettes to those that are color vision deficiency (CVD) safe. Default is FALSE, set to TRUE to enact filter.}

\item{BestFor}{Argument used to restrict displayed palettes to those that are best for either continuous or discrete data types. Arguments are "discrete" or "continuous"}
}
\value{
Graphic object.
}
\description{
Generates a graphic object with desired subset of palettes from WrensBookshelf.
}
\examples{
ShowBookshelf()
ShowBookshelf(CVDsafe = TRUE, BestFor = "continuous")
}
