% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowBook.R
\name{ShowBook}
\alias{ShowBook}
\title{View a book palette}
\usage{
ShowBook(name)
}
\arguments{
\item{name}{Name of desired palette.}
}
\value{
Graphic object.
}
\description{
Generates a graphic object using a specified palette from WrensBookshelf.
}
\examples{
ShowBook(name ="BabyWrenAndTheGreatGift")
}
