% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrensBookshelf.R
\docType{data}
\name{WrensBookshelf}
\alias{WrensBookshelf}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 40.
}
\usage{
WrensBookshelf
}
\value{
No return value, sets a list of palettes to be used with WrensBookshelf functions
}
\description{
Use names(WrensBookshelf) to display all possible palette names. Current available palettes:
\code{BabyBeluga}, \code{BabyWrenAndTheGreatGift}, \code{BlueberriesForSal}, \code{BoTheBrave},
\code{BoTheBraveCVDS}, \code{BurtDow}, \code{BurtDowCVDS}, \code{CapsForSale}, \code{GustavoTheShyGhostCover},
\code{GustavoTheShyGhostGraveyard}, \code{HaroldAndThePurpleCrayon}, \code{JeffGoesWild},
\code{JulienIsAMermaid}, \code{LittleBlueHouseBesideTheSeaCont}, \code{LittleBlueHouseBesideTheSeaDisc},
\code{Madelline}, \code{MakeWayForDucklings}, \code{MoreThanALittle}, \code{OhThePlacesYoullGo},
\code{Opposites}, \code{StackTheCats}, \code{TheDucklingGetsACookie}, \code{TheHoneybee},
\code{TheHoneybeeDiverging}, \code{ThePoutPoutFish}, \code{TheSnowyDay}, \code{TheStoryOfBabar},
\code{TheseWordsIShapedForYou}, \code{ThisMooseBelongsToMe}, \code{ThreeLittleVikings}, \code{TigerDays},
\code{TinyPerfectThings}, \code{Vampenguin}, \code{WhatWellBuild}, \code{WhereTheWildThingsAre}, and \code{YouMatter}.
}
\keyword{datasets}
